FatalError functional component + tsx
